Add left navigation to webapp
