+++ packages/webapp/src/App.tsx
@@ -1,7 +1,9 @@
 import { Routes, Route, Navigate } from 'react-router-dom';
 import Signup from './pages/Signup';
 import Signin from './pages/Signin';
-import Home from './pages/Home';
+import IntegrationList from './pages/IntegrationList';
+import ConnectionList from './pages/ConnectionList';
+import ProjectSettings from './pages/ProjectSettings';
 import PrivateRoute from './components/PrivateRoute';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
@@ -10,8 +12,15 @@ const App = () => {
     return (
         <>
             <Routes>
-                <Route path="/" element={<PrivateRoute />}>
-                    <Route path="/" element={<Home />} />
+                <Route path="/" element={<Navigate to="/integrations" replace />} />
+                <Route path="/integrations" element={<PrivateRoute />}>
+                    <Route path="/integrations" element={<IntegrationList />} />
+                </Route>
+                <Route path="/connections" element={<PrivateRoute />}>
+                    <Route path="/connections" element={<ConnectionList />} />
+                </Route>
+                <Route path="/project-settings" element={<PrivateRoute />}>
+                    <Route path="/project-settings" element={<ProjectSettings />} />
                 </Route>
                 <Route path="/signin" element={<Signin />} />
                 <Route path="/signup" element={<Signup />} />
