+++ packages/webapp/src/pages/ConnectionList.tsx
@@ -1,0 +1,14 @@
+import TopNavBar from '../components/TopNavBar';
+import LeftNavBar, { LeftNavBarItems } from '../components/LeftNavBar';
+
+export default function ProjectSettings() {
+    return (
+        <div className="h-full">
+            <TopNavBar />
+            <div className="flex h-full">
+                <LeftNavBar selectedItem={LeftNavBarItems.Connections} />
+                <div>Connections</div>
+            </div>
+        </div>
+    );
+}
