+++ packages/webapp/src/pages/IntegrationList.tsx
@@ -1,0 +1,14 @@
+import TopNavBar from '../components/TopNavBar';
+import LeftNavBar, { LeftNavBarItems } from '../components/LeftNavBar';
+
+export default function ProjectSettings() {
+    return (
+        <div className="h-full">
+            <TopNavBar />
+            <div className="flex h-full">
+                <LeftNavBar selectedItem={LeftNavBarItems.Integrations} />
+                <div>Integrations</div>
+            </div>
+        </div>
+    );
+}
